+++ src/multi_bot_agentic/safety.py
@@ -70,6 +70,7 @@ class SafetyPolicy:
             "iban_check",
             "ics_parse",
             "ini_parse",
+            "isbn13",
             "json_diff_paths",
             "json_flatten",
             "json_format",